Nuxt-migration: Removed register-service-worker and add nuxt/pwa with notification
