+++ layouts/default.vue
@@ -16,6 +16,7 @@
 
 <script lang="ts">
 import { Component, Vue } from 'nuxt-property-decorator'
+import { NotificationProgrammatic as Notification } from 'buefy'
 import { cryptoWaitReady } from '@polkadot/util-crypto'
 import keyring from '@polkadot/ui-keyring'
 import isShareMode from '@/utils/isShareMode'
@@ -82,6 +83,7 @@ export default class Dashboard extends Vue {
     this.mountWasmCrypto()
     this.fetchIndexer()
     this.getKsmPrice()
+    this.checkVersion()
   }
 
   private async fetchIndexer() {
@@ -120,6 +122,37 @@ export default class Dashboard extends Vue {
       this.$buefy.snackbar.open(type as any)
       // this.$router.push({ name: 'error' });
       console.warn('Do something', error)
+    }
+  }
+
+  private async checkVersion() {
+    //@ts-ignore
+    const workbox = await window.$workbox;
+    if (workbox) {
+      workbox.addEventListener('installed', (event) => {
+        console.log(
+          'App is being served from cache by a service worker.\n' +
+            'For more details, visit https://pwa.nuxtjs.org/'
+        )
+
+        if (event.isUpdate) {
+          console.log('New content is available; please refresh.')
+          const notif = Notification.open({
+            message: 'New version is ready. Close to upgrade.',
+            queue: false,
+            type: 'is-info is-dark',
+            position: 'is-top-left',
+            indefinite: true,
+            hasIcon: true,
+          })
+
+          notif.$on('close', () => {
+            window.sessionStorage.clear()
+            window.localStorage.clear()
+            window.location.reload()
+          })
+        }
+      });
     }
   }
